Move overlapping output detection to :snapshots
