+++ subprojects/core/src/test/groovy/org/gradle/api/internal/tasks/execution/DefaultTaskCacheabilityResolverTest.groovy
@@ -18,7 +18,6 @@ package org.gradle.api.internal.tasks.execution
 
 import com.google.common.collect.ImmutableSortedSet
 import org.gradle.api.GradleException
-import org.gradle.api.internal.OverlappingOutputs
 import org.gradle.api.internal.TaskInternal
 import org.gradle.api.internal.tasks.properties.CacheableOutputFilePropertySpec
 import org.gradle.api.internal.tasks.properties.OutputFilePropertySpec
@@ -26,6 +25,7 @@ import org.gradle.api.specs.Spec
 import org.gradle.internal.execution.caching.CachingDisabledReason
 import org.gradle.internal.execution.caching.CachingDisabledReasonCategory
 import org.gradle.internal.file.RelativeFilePathResolver
+import org.gradle.internal.fingerprint.OverlappingOutputs
 import spock.lang.Specification
 
 import javax.annotation.Nullable
