+++ subprojects/core/src/test/groovy/org/gradle/api/internal/tasks/execution/ResolveBeforeExecutionOutputsTaskExecuterTest.groovy
@@ -18,7 +18,6 @@ package org.gradle.api.internal.tasks.execution
 
 import com.google.common.collect.ImmutableSortedMap
 import com.google.common.collect.ImmutableSortedSet
-import org.gradle.api.internal.OverlappingOutputs
 import org.gradle.api.internal.TaskInternal
 import org.gradle.api.internal.tasks.TaskExecuter
 import org.gradle.api.internal.tasks.TaskExecuterResult
@@ -27,6 +26,7 @@ import org.gradle.api.internal.tasks.TaskStateInternal
 import org.gradle.api.internal.tasks.properties.OutputFilePropertySpec
 import org.gradle.api.internal.tasks.properties.TaskProperties
 import org.gradle.internal.execution.history.AfterPreviousExecutionState
+import org.gradle.internal.fingerprint.OverlappingOutputs
 import org.gradle.internal.fingerprint.impl.AbsolutePathFingerprintingStrategy
 import org.gradle.internal.snapshot.FileSystemSnapshot
 import spock.lang.Specification
